+++ cmd/siva/impl/pack_test.go
@@ -1,4 +1,4 @@
-package main
+package impl
 
 import (
 	"io/ioutil"
