+++ cmd/siva/impl/unpack_windows.go
@@ -1,5 +1,5 @@
 // +build windows
 
-package main
+package impl
 
 const defaultPerms = 0666
