+++ cmd/siva/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "gopkg.in/src-d/go-siva.v1/cmd/siva/impl"
+
+func main() {
+    impl.Main()
+}
